Feature: Center content and footer
- Centered the footer text.
- Centered the service cards vertically on the page.
