+++ global.css
@@ -83,13 +83,23 @@ body.dark {
   background: var(--clr-accent);
 }
 /* Cards Grid */
+body {
+  display: flex;
+  flex-direction: column;
+  justify-content: center;
+  align-items: center;
+}
+
+main {
+  width: 100%;
+}
+
 .grid-container {
   width: min(75rem, 100%);
   margin-inline: auto;
-  margin-top: 2.8rem;
   display: grid;
-  grid-template-columns: repeat(auto-fit, minmax(min(19rem, 100%), 1fr));
-  gap: 2.4rem;
+  grid-template-columns: repeat(auto-fit, minmax(min(15rem, 100%), 1fr));
+  gap: 1.5rem;
 }
 .card {
   padding: 2.3rem;
@@ -276,8 +286,8 @@ footer {
   background: var(--clr-footer);
   color: #fff;
   font-size: 1rem;
-  text-align: right;
-  padding: 1.1rem 2.5rem 1.1rem 0.8rem;
+  text-align: center;
+  padding: 1.1rem 0;
   position: fixed;
   left: 0;
   bottom: 0;
