+++ index.html
@@ -13,6 +13,7 @@
   <nav class="ops-nav">
     <span class="ops-logo">OPS</span>
     <div class="nav-links">
+      <a href="index.html" class="nav-link">Home</a>
       <a href="business.html" class="nav-link">Business Operations</a>
       <a href="contactcenter.html" class="nav-link">Contact Center</a>
       <a href="itsupport.html" class="nav-link">IT Support</a>
